Add MaxAmount guard in MPTIssuanceCreate

+++ src/test/jtx/mpt.h
@@ -38,7 +38,7 @@ create(jtx::Account const& account);
 Json::Value
 create(
     jtx::Account const& account,
-    std::uint32_t const maxAmt,
+    std::uint64_t const maxAmt,
     std::uint8_t const assetScale,
     std::uint16_t transferFee,
     std::string metadata);
